other admin CRUD, home page with manga scroller, ecc
Added AdminService, to separate the business logic with the UserService and added a custom exception "MangaNotFoundException". In the AdminService I have created some new crud methods for admin purposes.
Added a method in mangaSercive to pick up the top ten scored manga for the scroller on the home page.
Added the scroller on the home page. ecc...